Update module: Show test - score - infor user test

+++ src/component/Admin/class.managemnet/class.management.tsx
@@ -13,19 +13,15 @@ import TestRow from './class.data';
 const ClassManagement = () => {
     const [testData, setTestData] = useState([]);
     const [loading, setLoading] = useState(true);
-    const [openLearnMore, setOpenLearnMore] = useState(false);
-    const [dataLearnMore, setDataLearnMore] = useState({});
-    // const [formTestData, setFormTestData] = useState({
-    //     name: '',
-    //     type: 'Reading', 
-    //     description: '',
-    // });
+
+    const [dataShowMore, setDataShowMore] = useState([]);
+    const [dataWritingShowMore, setDataWritingShowMore] = useState([]);
 
     const fetchData = async () => {
         setLoading(true);
         const response = await fetch(`${process.env.NEXT_PUBLIC_STRAPI_LINK_API_URL}/tests?populate=*`, { cache: 'no-store' });
         const data = await response.json();
-        console.log(data);
+        // console.log(data);
 
         setTestData(data.data);
         setLoading(false);
@@ -38,27 +34,89 @@ const ClassManagement = () => {
 
 
     const handleDeleteTest = async (id: any) => {
-        await fetch(`${process.env.NEXT_PUBLIC_STRAPI_LINK_API_URL}/tests/${id}`, {
-            method: 'DELETE',
-        });
-        fetchData();
+
+
+        if (window.confirm('Bạn có muốn xoá bài test này khỏi hệ thống không?')) {
+            await fetch(`${process.env.NEXT_PUBLIC_STRAPI_LINK_API_URL}/tests/${id}`, {
+                method: 'DELETE',
+            });
+            fetchData();
+        };
+
     };
 
     const handleRefresh = () => {
         fetchData();
     };
-    console.log(testData);
-    console.log(loading);
 
 
-    const handleLearnMore = (test: any) => {
-        setDataLearnMore(test);
-        setOpenLearnMore(true);
-        alert("learn more")
-        console.log(test);
+    const handleShowMore = async (test: any) => {
+        //  console.log(test.attributes.type);
+
+        if (test.attributes.type === 'Wrting') {
+            //   console.log(test);
+
+            const writingId = test.attributes.wrting.data.id;
+            const response = await fetch(`${process.env.NEXT_PUBLIC_STRAPI_LINK_API_URL}/wrtings/${writingId}?populate=*`);
+            const data_writing = await response.json();
+
+            const updatedData: any = await Promise.all(data_writing.data.attributes.answer_wrtings.data.map(async (answer_wrting: any) => {
+                const userId = answer_wrting.id;
+                const res = await fetch(`${process.env.NEXT_PUBLIC_STRAPI_LINK_API_URL}/answer-wrtings/${userId}?populate=*`);
+                const userData = await res.json();
+
+                return {
+                    name: userData.data.attributes.users_permissions_user.data.attributes.fullname,
+                    answer_task1: userData.data.attributes.task1Answer,
+                    answer_task2: userData.data.attributes.task2Answer
+                };
+            }));
+
+            setDataWritingShowMore(updatedData);
+
+            const data_show_writing = {
+                title: test.attributes.name,
+                data: updatedData,
+                type: 'writing'
+            };
+
+            // console.log(data_show_writing);
+
+            const url = `/dashboard/show-more?data=${encodeURIComponent(JSON.stringify(data_show_writing))}`;
+            window.open(url, "_blank", "width=800,height=600");
 
 
-    }
+        } else {
+            //   console.log(test);
+
+            const data_student_tests = test.attributes.student_tests.data;
+            const updatedData: any = await Promise.all(data_student_tests.map(async (student_test: any) => {
+                const userId = student_test.id;
+                const res = await fetch(`${process.env.NEXT_PUBLIC_STRAPI_LINK_API_URL}/student-tests/${userId}?populate=*`);
+                const userData = await res.json();
+
+                return {
+                    name: userData.data.attributes.user?.data.attributes.fullname,
+                    score: userData.data.attributes.score,
+                    answer_data: userData.data.attributes.answer_data
+                };
+            }));
+
+            setDataShowMore(updatedData);
+
+            const data_show = {
+                title: test.attributes.name,
+                data: updatedData,
+                type: 'not_writing'
+            };
+
+            // console.log(data_show);
+
+            const url = `/dashboard/show-more?data=${encodeURIComponent(JSON.stringify(data_show))}`;
+            window.open(url, "_blank", "width=600,height=400");
+        }
+    };
+
     return (
         <Paper sx={{ maxWidth: 936, margin: 'auto', overflow: 'hidden' }}>
             <AppBar
@@ -84,9 +142,7 @@ const ClassManagement = () => {
                             />
                         </Grid>
                         <Grid item>
-                            {/* <Button variant="contained" sx={{ mr: 1 }} onClick={handleAddTest}>
-                                Add Test
-                            </Button> */}
+
                             <Tooltip title="Refresh">
                                 <IconButton onClick={handleRefresh}>
                                     <RefreshIcon color="inherit" sx={{ display: 'block' }} />
@@ -116,7 +172,7 @@ const ClassManagement = () => {
                                             <TestRow
                                                 key={index}
                                                 test={test}
-                                                handleLearnMore={handleLearnMore}
+                                                handleLearnMore={handleShowMore}
                                                 handleDeleteTest={handleDeleteTest}
                                             />
                                         </>
